Modificacion de toda la conexion

+++ bonita_smile_v1/Servicios/Carpeta_archivos.cs
@@ -16,10 +16,12 @@ namespace bonita_smile_v1.Servicios
         private MySqlConnection conexionBD;
         Conexion obj = new Conexion();
         Test_Internet ti = new Test_Internet();
+        private bool online;
 
-        public Carpeta_archivos()
+        public Carpeta_archivos(bool online)
         {
-            this.conexionBD = obj.conexion();
+            this.conexionBD = obj.conexion(online);
+            this.online = online;
         }
 
         public List<Carpeta_archivosModel> MostrarCarpeta_archivos()
@@ -55,19 +57,43 @@ namespace bonita_smile_v1.Servicios
 
         public bool eliminarCarpeta_archivos(string id_carpeta)
         {
-            query = "DELETE FROM carpeta_archivos where id_carpeta='" + id_carpeta+"'";
+            
+            bool internet = ti.Test();
             try
             {
-                conexionBD.Open();
-                MySqlCommand cmd = new MySqlCommand(query, conexionBD);
-                cmd.ExecuteReader();
-                conexionBD.Close();
-                return true;
 
+                MySqlCommand cmd; ;
+                if (online)
+                {
+                    if (!internet)
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ELIMINAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI NO LO HAY, ENTONCES NO HACER NADA Y SEGUIR MANTENIENDO QUERIES EN EL ARCHIVO 
+                    }
+                    else
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ELIMINAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI LO HAY, ENTONCES INSERTAR TODOS LOS QUERIES DEL ARCHIVO
+
+                        Sincronizar sincronizar = new Sincronizar();
+                        sincronizar.insertarArchivoEnServidor(conexionBD);
+                    }
+                }
+                else
+                {
+                    query = "DELETE FROM carpeta_archivos where id_carpeta='" + id_carpeta + "'";
+
+                    conexionBD.Open();
+                    cmd = new MySqlCommand(query, conexionBD);
+                    cmd.ExecuteReader();
+                    conexionBD.Close();
+
+                    Escribir_Archivo ea = new Escribir_Archivo();
+                    ea.escribir(query + ";");
+                }
+                return true;
             }
             catch (MySqlException ex)
             {
-                MessageBox.Show(ex.ToString());
+                System.Windows.MessageBox.Show(ex.ToString());
                 conexionBD.Close();
                 return false;
             }
@@ -78,36 +104,44 @@ namespace bonita_smile_v1.Servicios
             string auxiliar_identificador = "";
             Seguridad seguridad = new Seguridad();
             auxiliar_identificador = seguridad.SHA1(nombre_carpeta + id_paciente+DateTime.Now);
+            
             bool internet = ti.Test();
 
-            if (!internet)
-            {
-               
-                 
-                query = "INSERT INTO carpeta_archivos (id_carpeta,nombre_carpeta,id_paciente,auxiliar_identificador) VALUES('"+auxiliar_identificador +"','"+ nombre_carpeta + "','" + id_paciente + "','<!--" + auxiliar_identificador + "-->')";
-            }
-            else
-            {
-                
-                query = "INSERT INTO carpeta_archivos (id_carpeta,nombre_carpeta,id_paciente) VALUES('" + auxiliar_identificador + "','" + nombre_carpeta + "','" + id_paciente +"')";
-            }
             try
             {
-                conexionBD.Open();
-                MySqlCommand cmd = new MySqlCommand(query, conexionBD);
-                cmd.ExecuteReader();
-                if (!internet)
-                {
-                    Escribir_Archivo ea = new Escribir_Archivo();
-                    ea.escribir( query + ";");
-                }
-                conexionBD.Close();
-                return true;
 
+                MySqlCommand cmd; ;
+                if (online)
+                {
+                    if (!internet)
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA INSERTAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI NO LO HAY, ENTONCES NO HACER NADA Y SEGUIR MANTENIENDO QUERIES EN EL ARCHIVO 
+                    }
+                    else
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA INSERTAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI LO HAY, ENTONCES INSERTAR TODOS LOS QUERIES DEL ARCHIVO
+
+                        //query = "INSERT INTO usuario (id_usuario,alias,nombre,apellidos,password,id_rol) VALUES('" + auxiliar_identificador + "','" + alias + "','" + nombre + "','" + apellidos + "','" + password + "'," + id_rol + ")";
+                        Sincronizar sincronizar = new Sincronizar();
+                        sincronizar.insertarArchivoEnServidor(conexionBD);
+                    }
+                }
+                else
+                {
+                    query = "INSERT INTO carpeta_archivos (id_carpeta,nombre_carpeta,id_paciente,auxiliar_identificador) VALUES('" + auxiliar_identificador + "','" + nombre_carpeta + "','" + id_paciente + "','<!--" + auxiliar_identificador + "-->')";
+                    conexionBD.Open();
+                    cmd = new MySqlCommand(query, conexionBD);
+                    cmd.ExecuteReader();
+                    conexionBD.Close();
+
+                    Escribir_Archivo ea = new Escribir_Archivo();
+                    ea.escribir(query + ";");
+                }
+                return true;
             }
             catch (MySqlException ex)
             {
-                MessageBox.Show(ex.ToString());
+                System.Windows.MessageBox.Show(ex.ToString());
                 conexionBD.Close();
                 return false;
             }
@@ -115,35 +149,46 @@ namespace bonita_smile_v1.Servicios
 
         public bool actualizarCarpeta_archivos(string id_carpeta, string nombre_carpeta, string id_paciente)
         {
+            
+
             bool internet = ti.Test();
-            if (!internet)
-            {
-                //Seguridad seguridad = new Seguridad();
-                // = seguridad.Encriptar(nombre_carpeta + id_paciente);
-                string auxiliar_identificador = MostrarCarpeta_Archivos_Update(id_carpeta);
-                query = "UPDATE carpeta_archivos set nombre_carpeta = '" + nombre_carpeta + "',id_paciente = '" + id_paciente + "',auxiliar_identificador = '" + auxiliar_identificador + "' where id_carpeta = '" + id_carpeta+"'";
-            }
-            else
-            {
-                query = "UPDATE carpeta_archivos set nombre_carpeta = '" + nombre_carpeta + "',id_paciente = '" + id_paciente + "' where id_carpeta = '" + id_carpeta+"'";
-            }
             try
             {
-                conexionBD.Open();
-                MySqlCommand cmd = new MySqlCommand(query, conexionBD);
-                cmd.ExecuteReader();
-                conexionBD.Close();
-                return true;
 
+                MySqlCommand cmd; ;
+                if (online)
+                {
+                    if (!internet)
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ACTUALIZAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI NO LO HAY, ENTONCES NO HACER NADA Y SEGUIR MANTENIENDO QUERIES EN EL ARCHIVO 
+                    }
+                    else
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA INSEACTUALIZARRTAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI LO HAY, ENTONCES INSERTAR TODOS LOS QUERIES DEL ARCHIVO
+
+                        //query = "UPDATE usuario set alias = '" + alias + "',nombre = '" + nombre + "',apellidos = '" + apellidos + "',password = '" + password + "',id_rol = " + id_rol + " where id_usuario = '" + id_usuario + "'";
+                        Sincronizar sincronizar = new Sincronizar();
+                        sincronizar.insertarArchivoEnServidor(conexionBD);
+                    }
+                }
+                else
+                {
+                    //string auxiliar_identificador = MostrarUsuario_Update(id_usuario);
+                    query = "UPDATE carpeta_archivos set nombre_carpeta = '" + nombre_carpeta + "',id_paciente = '" + id_paciente + "',auxiliar_identificador = '" + id_carpeta + "' where id_carpeta = '" + id_carpeta + "'";
+
+                    conexionBD.Open();
+                    cmd = new MySqlCommand(query, conexionBD);
+                    cmd.ExecuteReader();
+                    conexionBD.Close();
+
+                    Escribir_Archivo ea = new Escribir_Archivo();
+                    ea.escribir(query + ";");
+                }
+                return true;
             }
             catch (MySqlException ex)
             {
-                MessageBox.Show(ex.ToString());
-                if (!internet)
-                {
-                    Escribir_Archivo ea = new Escribir_Archivo();
-                    ea.escribir( query + ";");
-                }
+                System.Windows.MessageBox.Show(ex.ToString());
                 conexionBD.Close();
                 return false;
             }
