+++ bonita_smile_v1/Servicios/Foto_estudio_carpeta.cs
@@ -20,17 +20,19 @@ namespace bonita_smile_v1.Servicios
         private MySqlConnection conexionBD;
         Conexion obj = new Conexion();
         Test_Internet ti = new Test_Internet();
+        private bool online;
 
-        public Fotos_estudio_carpeta()
+        public Fotos_estudio_carpeta(bool online)
         {
-            this.conexionBD = obj.conexion();
+            this.conexionBD = obj.conexion(online);
+            this.online = online;
         }
 
         public List<Fotos_estudio_carpetaModel> MostrarFoto_estudio_carpeta(string id_carpeta, string id_paciente)
         {
             List<Fotos_estudio_carpetaModel> listaFoto_estudio_carpeta = new List<Fotos_estudio_carpetaModel>();
             query = "SELECT  * FROM fotos_estudio_carpeta where id_carpeta='" + id_carpeta + "' and id_paciente='" + id_paciente+"'";
-
+            string foto_recortada= "";
             try
             {
                 conexionBD.Open();
@@ -45,7 +47,10 @@ namespace bonita_smile_v1.Servicios
                     fotos_Estudio_CarpetaModel.id_foto = reader[0].ToString();
                     fotos_Estudio_CarpetaModel.id_carpeta = reader[1].ToString();
                     fotos_Estudio_CarpetaModel.id_paciente = reader[2].ToString();
-                    fotos_Estudio_CarpetaModel.foto = reader[3].ToString();
+                    foto_recortada= reader[3].ToString();
+                    foto_recortada = foto_recortada.Substring(35);
+                    fotos_Estudio_CarpetaModel.foto = foto_recortada;
+                    fotos_Estudio_CarpetaModel.foto_completa= reader[3].ToString();
                     fotos_Estudio_CarpetaModel.imagen = LoadImage(@"C:\bs\" + reader[3].ToString());
 
                     listaFoto_estudio_carpeta.Add(fotos_Estudio_CarpetaModel);
@@ -62,23 +67,44 @@ namespace bonita_smile_v1.Servicios
 
         public bool eliminarFoto_estudio_carpeta(string id_foto)
         {
-            query = "DELETE FROM fotos_estudio_carpeta where id_foto='" + id_foto+"'";
+            
+
+            bool internet = ti.Test();
             try
             {
-                conexionBD.Open();
-                MySqlCommand cmd = new MySqlCommand(query, conexionBD);
-                cmd.ExecuteReader();
-                conexionBD.Close();
-                if (!ti.Test())
+
+                MySqlCommand cmd; ;
+                if (online)
                 {
-                   // Escribir_Archivo ea = new Escribir_Archivo();
-                   // ea.escribir(@"c:\offline\script_temporal.txt", query + ";");
+                    if (!internet)
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ELIMINAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI NO LO HAY, ENTONCES NO HACER NADA Y SEGUIR MANTENIENDO QUERIES EN EL ARCHIVO 
+                    }
+                    else
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ELIMINAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI LO HAY, ENTONCES INSERTAR TODOS LOS QUERIES DEL ARCHIVO
+
+                        Sincronizar sincronizar = new Sincronizar();
+                        sincronizar.insertarArchivoEnServidor(conexionBD);
+                    }
+                }
+                else
+                {
+                    query = "DELETE FROM fotos_estudio_carpeta where id_foto='" + id_foto + "'";
+
+                    conexionBD.Open();
+                    cmd = new MySqlCommand(query, conexionBD);
+                    cmd.ExecuteReader();
+                    conexionBD.Close();
+
+                    Escribir_Archivo ea = new Escribir_Archivo();
+                    ea.escribir(query + ";");
                 }
                 return true;
             }
             catch (MySqlException ex)
             {
-                MessageBox.Show(ex.ToString());
+                System.Windows.MessageBox.Show(ex.ToString());
                 conexionBD.Close();
                 return false;
             }
@@ -88,81 +114,91 @@ namespace bonita_smile_v1.Servicios
         {
             string auxiliar_identificador = "";
             Seguridad seguridad = new Seguridad();
-            MessageBox.Show((foto.Length/2)+"");
-            MessageBox.Show(foto.Length + "");
-            foto = foto.Substring((foto.Length/4)*3);
-
-
-            MessageBox.Show("foto = " + foto);
-            //string auxiliar_carpeta = seguridad.Encriptar(id_carpeta);
-           //string  auxiliar_paciente = seguridad.Encriptar(id_paciente);
-           //string  auxiliar_foto = seguridad.Encriptar(foto);
-            auxiliar_identificador=seguridad.SHA1(id_carpeta+id_paciente+foto);
-            //auxiliar_identificador = seguridad.Encriptar(id_carpeta + id_paciente + foto);
+            auxiliar_identificador = seguridad.SHA1(id_carpeta + id_paciente + foto);
             bool internet = ti.Test();
-            if (!internet)
-            {
-               
-                query = "INSERT INTO fotos_estudio_carpeta (id_foto,id_carpeta,id_paciente,foto,auxiliar_identificador) VALUES('"+auxiliar_identificador +"','"+ id_carpeta + "','" + id_paciente + "','" + foto + "','<!--" + auxiliar_identificador + "-->')";
-            }
-            else
-            {
-                query = "INSERT INTO fotos_estudio_carpeta (id_foto,id_carpeta,id_paciente,foto) VALUES('" + auxiliar_identificador + "','" + id_carpeta + "','" + id_paciente + "','" + foto + "')";
-            }
-
             try
             {
-                conexionBD.Open();
-                MySqlCommand cmd = new MySqlCommand(query, conexionBD);
-                cmd.ExecuteReader();
-                conexionBD.Close();
-                if (!internet)
+
+                MySqlCommand cmd; ;
+                if (online)
                 {
+                    if (!internet)
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA INSERTAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI NO LO HAY, ENTONCES NO HACER NADA Y SEGUIR MANTENIENDO QUERIES EN EL ARCHIVO 
+                    }
+                    else
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA INSERTAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI LO HAY, ENTONCES INSERTAR TODOS LOS QUERIES DEL ARCHIVO
+
+                        //query = "INSERT INTO usuario (id_usuario,alias,nombre,apellidos,password,id_rol) VALUES('" + auxiliar_identificador + "','" + alias + "','" + nombre + "','" + apellidos + "','" + password + "'," + id_rol + ")";
+                        Sincronizar sincronizar = new Sincronizar();
+                        sincronizar.insertarArchivoEnServidor(conexionBD);
+                    }
+                }
+                else
+                {
+                    query = "INSERT INTO fotos_estudio_carpeta (id_foto,id_carpeta,id_paciente,foto,auxiliar_identificador) VALUES('" + auxiliar_identificador + "','" + id_carpeta + "','" + id_paciente + "','" + foto + "','<!--" + auxiliar_identificador + "-->')";
+
+                    conexionBD.Open();
+                    cmd = new MySqlCommand(query, conexionBD);
+                    cmd.ExecuteReader();
+                    conexionBD.Close();
+
                     Escribir_Archivo ea = new Escribir_Archivo();
-                    ea.escribir( query + ";");
+                    ea.escribir(query + ";");
                 }
                 return true;
-
             }
             catch (MySqlException ex)
             {
-                MessageBox.Show(ex.ToString());
+                System.Windows.MessageBox.Show(ex.ToString());
                 conexionBD.Close();
                 return false;
             }
         }
 
+
         public bool actualizarFoto_estudio_carpeta(string id_foto, string id_carpeta, string id_paciente, string foto)
         {
+
             bool internet = ti.Test();
-            if (!internet)
-            {
-                //Seguridad seguridad = new Seguridad();
-                // = seguridad.Encriptar(id_carpeta + id_paciente + foto);
-                string auxiliar_identificador = MostrarFotos_Update(id_foto);
-                query = "UPDATE fotos_estudio_carpeta set id_paciente = '" + id_paciente + "',id_carpeta = '" + id_carpeta + "',foto = '" + foto + ",auxiliar_identificador = '" + auxiliar_identificador + "' where id_foto = '" + id_foto+"'";
-            }
-            else
-            {
-                query = "UPDATE fotos_estudio_carpeta set id_paciente = '" + id_paciente + "',id_carpeta = '" + id_carpeta + "',foto = '" + foto + "' where id_foto = '" + id_foto+"'";
-            }
             try
             {
-                conexionBD.Open();
-                MySqlCommand cmd = new MySqlCommand(query, conexionBD);
-                cmd.ExecuteReader();
-                conexionBD.Close();
-                if (!ti.Test())
+
+                MySqlCommand cmd; ;
+                if (online)
                 {
+                    if (!internet)
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA ACTUALIZAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI NO LO HAY, ENTONCES NO HACER NADA Y SEGUIR MANTENIENDO QUERIES EN EL ARCHIVO 
+                    }
+                    else
+                    {
+                        //EN CASO DE REALIZAR UNA PETICION PARA INSEACTUALIZARRTAR EN SERVIDOR VERIFICAR SI HAY INTERNET, SI LO HAY, ENTONCES INSERTAR TODOS LOS QUERIES DEL ARCHIVO
+
+                        //query = "UPDATE usuario set alias = '" + alias + "',nombre = '" + nombre + "',apellidos = '" + apellidos + "',password = '" + password + "',id_rol = " + id_rol + " where id_usuario = '" + id_usuario + "'";
+                        Sincronizar sincronizar = new Sincronizar();
+                        sincronizar.insertarArchivoEnServidor(conexionBD);
+                    }
+                }
+                else
+                {
+                    //string auxiliar_identificador = MostrarUsuario_Update(id_usuario);
+                    query = "UPDATE fotos_estudio_carpeta set id_paciente = '" + id_paciente + "',id_carpeta = '" + id_carpeta + "',foto = '" + foto + ",auxiliar_identificador = '" + id_foto + "' where id_foto = '" + id_foto + "'";
+
+                    conexionBD.Open();
+                    cmd = new MySqlCommand(query, conexionBD);
+                    cmd.ExecuteReader();
+                    conexionBD.Close();
+
                     Escribir_Archivo ea = new Escribir_Archivo();
                     ea.escribir(query + ";");
                 }
                 return true;
-
             }
             catch (MySqlException ex)
             {
-                MessageBox.Show(ex.ToString());
+                System.Windows.MessageBox.Show(ex.ToString());
                 conexionBD.Close();
                 return false;
             }
@@ -205,12 +241,12 @@ namespace bonita_smile_v1.Servicios
             }
             else
             {
-                MessageBox.Show("No la encontro");
+                MessageBox.Show("No la encontro "+filename);
                 bi = new BitmapImage(new Uri(@"C:\bs\img1.jpg"));
             }
             return bi;
         }
-        public void fotos(string id_carpeta, string id_paciente)
+        /*public void fotos(string id_carpeta, string id_paciente)
         {
             string fotito = "";
             List<string> listaFoto_estudio_carpeta = new List<string>();
@@ -254,7 +290,7 @@ namespace bonita_smile_v1.Servicios
             conexionBD.Close();
             //return listaFoto_estudio_carpeta;
 
-        }
+        }*/
 
         public bool downloadFile(string servidor, string usuario, string password, string archivoOrigen, string carpetaDestino, int bufferdes)
         {
